chore: added comments for readability, renamed file to ensure name consistency
